+++ scripts/database-validator/internal/checkvalid/checkvalid.go
