+++ scripts/git-platforms-enumerator/main.go
