+++ scripts/git-relationship-generator/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"log"
 
-	"github.com/HUSTSecLab/OpenSift/cmd/git-relationship-generator/internal/pkgdep2git"
+	"github.com/HUSTSecLab/OpenSift/scripts/git-relationship-generator/internal/pkgdep2git"
 	"github.com/HUSTSecLab/OpenSift/pkg/config"
 	"github.com/HUSTSecLab/OpenSift/pkg/storage"
 	"github.com/spf13/pflag"
